requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1
django-cors-headers==3.14.0
djangorestframework==3.14.0
mysqlclient==2.1.1
pytz==2023.3
sqlparse==0.4.3
